+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support `fill0`, `fill1`, `fillx`, and `fillz`.
 - Support `rangeTo` operator.
 - Support string concatenation.
+- Support associative array `set` and `get`.
 
 ## [0.1.14]
 ### Importer